+++ src/types/configuration.ts
@@ -26,9 +26,7 @@ export interface ConfigurationClusterDefinition {
     port: number
   }>
   securityGroups: string[]
-  publicSubnets?: string[] // @deprecated: Please use subnets instead
-  privateSubnets?: string[] // @deprecated: Please use subnets instead
-  subnets?: string[] // TODO: make this required once other options are removed
+  subnets: string[]
   secrets?: {
     [name: string]: string
   }