+++ CHANGELOG.md
@@ -14,6 +14,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ps` command shows a list of services and tasks currently active in a cluster
 
 
+### Removed
+
+- No longer assign public IP address or allow public/private subnet config
+
+
 
 ## [0.5.0] - 2021-06-16